Move: 서버 API를 호출할 fetchMyOrder는 api 폴더로 옮기고 use 커스텀 훅이 아닌 parseOrder는 사용하는 파일 내로 이동
